requirements.txt
nltk==3.2.5
numpy==1.14.1
bllipparser==2016.9.11
scipy==1.0.0
newspaper3k==0.2.6
scikit-learn==0.19.1 
sklearn==0.0
tensorflow==1.6.0
autopep8==1.3.4
pylint==1.8.3
watson-developer-cloud==1.0.0
requests==2.18.4
Flask==1.0